+++ src/Traits/TypeValueHelpersTrait.php
@@ -39,6 +39,7 @@ use YdbPlatform\Ydb\Types\DecimalType;
 use YdbPlatform\Ydb\Types\DatetimeType;
 use YdbPlatform\Ydb\Types\TimestampType;
 use YdbPlatform\Ydb\Contracts\TypeContract;
+use YdbPlatform\Ydb\Types\YsonType;
 
 trait TypeValueHelpersTrait
 {
@@ -142,7 +143,7 @@ trait TypeValueHelpersTrait
             case 'STRING': return new StringType($value);
             case 'TEXT':
             case 'UTF8': return new Utf8Type($value);
-            // case 'YSON': return new StringValue(); // not implemented
+            case 'YSON': return new YsonType($value); // not implemented
             case 'JSON': return new JsonType($value);
             case 'UUID': return new StringType($value);
         }