+++ CHANGELOG.md
@@ -1,4 +1,6 @@
 * added transaction mode for retryTransaction
+* fix keepInCache param in YdbQuery
+* added Yson type
 * add logger as Ydb config
 * added snapshot mode in noninteractive transaction